requirements.txt
asgiref==3.8.1
black==24.8.0
click==8.1.7
colorama==0.4.6
Django==5.1.1
django-tenants==3.7.0
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
ipython==7.9.0
ipython-genutils==0.2.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
python-decouple==3.8
psycopg2-binary==2.9.9
PyJWT==2.9.0
sqlparse==0.5.1
tzdata==2024.2
nltk==3.9.1
openai==1.108.1